refactor symbolic mpo
